chore: merge fsContentQueries.ts and dbContentQueries.ts
