+++ src/app/feed.xml/route.ts
@@ -1,6 +1,6 @@
 import Rss from "rss";
 import { siteConfig } from "@/config/site";
-import { getPosts } from "@/lib/dbContentQueries";
+import { getPosts } from "@/lib/content-queries";
 
 export const dynamic = "force-static";
