+++ src/app/page.tsx
@@ -12,7 +12,7 @@ import { ViewDisplay } from "@/components/view-display";
 import { components, options } from "@/config/mdx-settings";
 import { siteConfig } from "@/config/site";
 import type { Post } from "@/config/types";
-import { getPosts } from "@/lib/dbContentQueries";
+import { getPosts } from "@/lib/content-queries";
 import { getPlaceholder } from "@/lib/images";
 import { cn } from "@/lib/utils";
